Added base EventStoreDB client (EventStoreDBClientBase) Added EventStoreDBProjectionManagementClient for projection management
